requirements.txt
fastapi
uvicorn
python-dotenv
twilio
requests

